Add files via upload
Violent crimes and other type of crimes raise broad concerns of residents in a city or a community. People move around partially due to public safety. If we could identify factors that highly related to crime rates, it will be useful for policy makers to take possible actions to reduce the crime rate. Also if we could predict crime rate well for a given city it will be helpful for people to make decision as where to live safely. Data is from 1995 Florida.